--- src/main/java/com/punnyajoshi/virtuallibrarysystem/commands/Command.java
@@ -1,7 +1,0 @@
-package com.khushiagrawall.virtuallibrarysystem.commands;
-
-import com.khushiagrawall.virtuallibrarysystem.models.Library;
-
-public interface Command {
-    void execute(Library lib);
-}
