--- src/main/java/com/punnyajoshi/virtuallibrarysystem/utlis/ScannerUtils.java
@@ -1,12 +1,0 @@
-package com.khushiagrawall.virtuallibrarysystem.utlis;
-
-import java.util.Scanner;
-
-public class ScannerUtils {
-    private static final Scanner sc=new Scanner(System.in);
-
-    public static int getIntInput(String message){
-        System.out.print(message+ " ");
-        return sc.nextInt();
-    }
-}
