+++ src/main/java/com/punnyajoshi/virtuallibrarysystemm/commands/BorrowBookCommand.java
@@ -1,0 +1,10 @@
+package com.punnyajoshi.virtuallibrarysystemm.commands;
+
+import com.punnyajoshi.virtuallibrarysystemm.models.Library;
+
+public class BorrowBookCommand implements Command {
+    @Override
+    public void execute(Library lib) {
+        lib.borrowBook();
+    }
+}
