+++ src/main/java/com/punnyajoshi/virtuallibrarysystemm/commands/Command.java
@@ -1,0 +1,8 @@
+package com.punnyajoshi.virtuallibrarysystemm.commands;
+
+import com.punnyajoshi.virtuallibrarysystemm.models.Library;
+import com.punnyajoshi.virtuallibrarysystemm.commands.Command;
+
+public interface Command {
+    void execute(Library lib);
+}
