+++ src/main/java/com/punnyajoshi/virtuallibrarysystemm/utlis/ScannerUtils.java
@@ -1,0 +1,12 @@
+package com.punnyajoshi.virtuallibrarysystemm.utlis;
+
+import java.util.Scanner;
+
+public class ScannerUtils {
+    private static final Scanner sc=new Scanner(System.in);
+
+    public static int getIntInput(String message){
+        System.out.print(message+ " ");
+        return sc.nextInt();
+    }
+}
