--- src/test/java/com/punnyajoshi/virtuallibrarysystem/MainTest.java
@@ -1,4 +1,0 @@
-package com.khushiagrawall.virtuallibrarysystem;
-
-public class MainTest {
-}
