+++ src/test/java/com/punnyajoshi/virtuallibrarysystemm/MainTest.java
@@ -1,0 +1,4 @@
+package com.punnyajoshi.virtuallibrarysystemm;
+
+public class MainTest {
+}
